+++ .gitignore
@@ -18,3 +18,4 @@
 data/
 Tree.xml
 config.production.json
+config.json
