+++ web/context/context.go
@@ -1,7 +1,8 @@
-package web
+package context
 
 import (
 	"context"
+	"fmt"
 	"net/http"
 )
 
@@ -63,6 +64,7 @@ func SetError(r *http.Request, err error) {
 
 // SetResponse устанавливает в контекст json ответ
 func SetResponse(r *http.Request, data interface{}) {
+	fmt.Println(data)
 	*r = *r.WithContext(
 		WithResponseContext(
 			r.Context(),
