+++ web/getTree.go
@@ -2,7 +2,7 @@ package web
 
 import (
 	"EcoPasport/model"
-	context "EcoPasport/web/context"
+	"EcoPasport/web/context"
 	"fmt"
 	"net/http"
 )
@@ -10,9 +10,8 @@ import (
 func webGetTree(w http.ResponseWriter, r *http.Request) {
 	data, err := model.GetTree()
 	if err != nil {
-		context.SetError(r, fmt.Errorf("error get tree: %v", err))
+		context.SetError(r, fmt.Errorf("error get tree: %v", "t"))
 		return
 	}
-
 	context.SetResponse(r, data)
 }
